+++ internal/api/tasks.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"encoding/json"
 	"go_final_project/internal/storage"
 
 	"net/http"
@@ -12,8 +11,6 @@ type TasksResp struct {
 }
 
 func getTasksHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-
 	maxCount := 50
 	tasks, err := storage.Tasks(maxCount)
 	if err != nil {
@@ -24,19 +21,4 @@ func getTasksHandler(w http.ResponseWriter, r *http.Request) {
 	writeJson(w, TasksResp{
 		Tasks: tasks,
 	})
-}
-
-func writeError(w http.ResponseWriter, errorMsg string) {
-	w.WriteHeader(http.StatusInternalServerError)
-	json.NewEncoder(w).Encode(struct {
-		Error string `json:"error"`
-	}{
-		Error: errorMsg,
-	})
-}
-
-func writeJson(w http.ResponseWriter, data interface{}) {
-	if err := json.NewEncoder(w).Encode(data); err != nil {
-		http.Error(w, "ошибка формирования JSON", http.StatusInternalServerError)
-	}
 }
